+++ admin/src/scenes/dashboard-head-center/center/index.js
@@ -1,5 +1,4 @@
 import React from "react";
-import styled from "styled-components";
 import { Col, Row } from "reactstrap";
 
 import Status from "./status";
@@ -9,18 +8,10 @@ export default function Index() {
     <>
       <Row style={{}}>
         <Col md={12}>
-          <Title>Centre</Title>
+          <h2 className="m-0 font-bold text-2xl">Centre</h2>
         </Col>
       </Row>
       <Status />
     </>
   );
 }
-
-// Title line with filters
-const Title = styled.h2`
-  color: #242526;
-  font-weight: bold;
-  font-size: 28px;
-  margin-bottom: 10px;
-`;