+++ admin/src/scenes/dashboard/missions/ProposedPlaces.js
@@ -1,13 +1,13 @@
 import React from "react";
 import { Col, Row } from "reactstrap";
 import { Link } from "react-router-dom";
-import { Card, CardTitle, CardValueWrapper, CardValue, Subtitle } from "../../../components/dashboard";
+import { Card, CardTitle, CardValueWrapper, CardValue } from "../../../components/dashboard";
 import { colors } from "../../../utils";
 
 export default function ProposedPlaces({ missionPlaceLeft, missionPlaceTotal, filter, getLink }) {
   return (
     <React.Fragment>
-      <Subtitle>Places proposées par les structures</Subtitle>
+      <h3 className="mt-4 mb-2 text-xl">Places proposées par les structures</h3>
       <Row>
         <Col md={6} xl={4} key="total">
           <Link to={getLink({ base: "/mission", filter })}>