+++ admin/src/scenes/structure/index.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
+import useDocumentTitle from "../../hooks/useDocumentTitle";
 
 import View from "./view";
 import Edit from "./edit";
@@ -7,6 +8,8 @@ import List from "./list";
 import Create from "./create";
 
 export default function Index() {
+  useDocumentTitle("Structures");
+
   return (
     <Switch>
       <Route path="/structure/create" component={Create} />