+++ app/webpack.dev.config.js
@@ -36,7 +36,7 @@ module.exports = (env) => {
       rules: [
         {
           test: /\.css$/,
-          use: ["style-loader", "css-loader"],
+          use: ["style-loader", "css-loader", "postcss-loader"],
         },
         {
           test: /\.js$/,